pkg/sdk: combining all pkgs under `pkg/sdk` into a single pkg
