+++ pkg/sdk/action.go
@@ -12,14 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package action
+package sdk
 
 import (
 	"fmt"
 
 	"github.com/operator-framework/operator-sdk/pkg/k8sclient"
-	sdkTypes "github.com/operator-framework/operator-sdk/pkg/sdk/types"
 	"github.com/operator-framework/operator-sdk/pkg/util/k8sutil"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 // Create creates the provided object on the server and updates the arg
@@ -27,7 +27,7 @@ import (
 // Returns an error if the object’s TypeMeta(Kind, APIVersion) or ObjectMeta(Name/GenerateName, Namespace) is missing or incorrect.
 // Can also return an api error from the server
 // e.g AlreadyExists https://github.com/kubernetes/apimachinery/blob/master/pkg/api/errors/errors.go#L423
-func Create(object sdkTypes.Object) (err error) {
+func Create(object Object) (err error) {
 	_, namespace, err := k8sutil.GetNameAndNamespace(object)
 	if err != nil {
 		return err
@@ -59,7 +59,7 @@ func Create(object sdkTypes.Object) (err error) {
 // Returns an error if the object’s TypeMeta(Kind, APIVersion) or ObjectMeta(Name, Namespace) is missing or incorrect.
 // Can also return an api error from the server
 // e.g Conflict https://github.com/kubernetes/apimachinery/blob/master/pkg/api/errors/errors.go#L428
-func Update(object sdkTypes.Object) (err error) {
+func Update(object Object) (err error) {
 	_, namespace, err := k8sutil.GetNameAndNamespace(object)
 	if err != nil {
 		return err
@@ -86,12 +86,45 @@ func Update(object sdkTypes.Object) (err error) {
 	return nil
 }
 
+// DeleteOp wraps all the options for Delete().
+type DeleteOp struct {
+	metaDeleteOptions *metav1.DeleteOptions
+}
+
+// DeleteOption configures DeleteOp.
+type DeleteOption func(*DeleteOp)
+
+func NewDeleteOp() *DeleteOp {
+	op := &DeleteOp{}
+	op.setDefaults()
+	return op
+}
+
+func (op *DeleteOp) applyOpts(opts []DeleteOption) {
+	for _, opt := range opts {
+		opt(op)
+	}
+}
+
+func (op *DeleteOp) setDefaults() {
+	if op.metaDeleteOptions == nil {
+		op.metaDeleteOptions = &metav1.DeleteOptions{}
+	}
+}
+
+// WithDeleteOptions sets the metav1.DeleteOptions for the Delete() operation.
+func WithDeleteOptions(metaDeleteOptions *metav1.DeleteOptions) DeleteOption {
+	return func(op *DeleteOp) {
+		op.metaDeleteOptions = metaDeleteOptions
+	}
+}
+
 // Delete deletes the specified object
 // Returns an error if the object’s TypeMeta(Kind, APIVersion) or ObjectMeta(Name, Namespace) is missing or incorrect.
 // e.g NotFound https://github.com/kubernetes/apimachinery/blob/master/pkg/api/errors/errors.go#L418
 // “opts” configures the DeleteOptions
 // When passed WithDeleteOptions(o), the specified metav1.DeleteOptions are set.
-func Delete(object sdkTypes.Object, opts ...DeleteOption) (err error) {
+func Delete(object Object, opts ...DeleteOption) (err error) {
 	name, namespace, err := k8sutil.GetNameAndNamespace(object)
 	if err != nil {
 		return err
