--- pkg/sdk/handler/handler.go
@@ -1,31 +1,0 @@
-// Copyright 2018 The Operator-SDK Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package handler
-
-import (
-	sdkTypes "github.com/operator-framework/operator-sdk/pkg/sdk/types"
-)
-
-// Handler reacts to events and outputs actions.
-// If any intended action failed, the event would be re-triggered.
-// For actions done before the failed action, there is no rollback.
-type Handler interface {
-	Handle(sdkTypes.Context, sdkTypes.Event) error
-}
-
-var (
-	// RegisteredHandler is the user registered handler set by sdk.Handle()
-	RegisteredHandler Handler
-)
