+++ pkg/sdk/query-opt.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package query
+package sdk
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
